[oss] Open Sourcing of Compiler Representations (#91)
* Modified files

* Add some IR-specific files

* additional files

* Remove IR extraction files

* Restore some files

* working help for train and preprocess

* Remove irrelevant files

* lean up IR full file dataset mode

* Relaunch CI

* Edit circleci

* Another try for the CI

* set click verison to avoid black error

* remove mypy, black some files

* fix some ci errors

* Fix more CI

* further fix ci

* Add dataloaders utils

* update python processor in test

* typo

* removing obsolete test runners

* Restore test_pipeline

* try to fix test pipeline

* tedt_pipeline again

* Rename test files

* Add license and remove useless files

+++ codegen_sources/preprocessing/tests/pipeline/test_pipeline.py
@@ -15,16 +15,14 @@ import pytest
 
 
 from codegen_sources.preprocessing.preprocess import preprocess
+from codegen_sources.model.src.utils import AttrDict
+
 
 input_path = Path(__file__).parents[4].joinpath("data/test_dataset")
 bpe_path = Path(__file__).parents[4].joinpath("data/bpe/cpp-java-python")
 logger = logging.getLogger(__name__)
 
-
-class AttrDict(dict):
-    def __init__(self, *args, **kwargs) -> None:
-        super(AttrDict, self).__init__(*args, **kwargs)
-        self.__dict__ = self
+RELATIVE_TOLERANCE = 0.005
 
 
 def _deactivate_in_ci() -> None:
@@ -50,24 +48,52 @@ DEFAULT_PARAMETERS = AttrDict(
         "bpe_timeout": 2,
         "train_bpe_timeout": 5,
         "repo_split": True,
+        "fastbpe_code_path": None,
+        "fastbpe_vocab_path": None,
+        "fastbpe_use_vocab": False,
+        "tokenize_line_timeout": 60,
     }
 )
 
 
 @pytest.fixture(autouse=True)
-def setup(tmpdir):
-    if (input_path / "log").is_dir():
-        shutil.rmtree(input_path.joinpath("log"))
+def setup() -> None:
+    subdirs = ["log", "XLM-syml"]
+    for name in subdirs:
+        subpath = input_path / name
+        if subpath.is_dir():
+            try:
+                shutil.rmtree(subpath)
+            except OSError as e:
+                logger.warning(f"Could not delete the folder ({subpath}):\n{e}")
     for f in input_path.glob("*"):
-        if not f.name.endswith(".json.gz"):
-            f.unlink()
+        if not f.name.endswith(".json.gz") and f.name not in subdirs:
+            try:
+                f.unlink()
+            except OSError as e:
+                if not f.name.startswith(".nfs"):
+                    raise e
+                logger.warning(f"Could not delete file:\n{e}")
+
+
+def check_number_lines_is(nb):
+    assert abs(count_bpe_lines() - nb) / nb < RELATIVE_TOLERANCE, count_bpe_lines()
+
+
+def count_bpe_lines():
+    tok_files = list(input_path.glob("*.bpe")) + list(input_path.glob("*.bperob"))
+    count = 0
+    for path in tok_files:
+        with open(path) as f:
+            count += len(f.readlines())
+    return count
 
 
 # Roberta Mode
 def test_obfuscation_roberta_pipeline():
     args = AttrDict(DEFAULT_PARAMETERS)
     args.update(
-        {"langs": ["java", "python"], "mode": "obfuscation", "bpe_mode": "roberta",}
+        {"langs": ["java", "python"], "mode": "obfuscation", "bpe_mode": "roberta"}
     )
     _deactivate_in_ci()
     preprocess(args)
